+++ code-gen/upgrade-cluster-state-wrapper.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# If VERBOSE is true, then output line-by-line execution
+"${VERBOSE:-false}" && set -x
+"${EXIT_ON_FAILURE:-true}" && set -e
+
 # This script is a wrapper for the upgrade-cluster-state-repo.sh script and may be used to aid the operator in upgrading the
 # cluster state repo to a target Beluga version. It abstracts away the location of the upgrade-cluster-state-repo.sh, which
 # performs the actual cluster-state migration to a target Beluga version. The script must be run from the root of the