+++ code-gen/upgrade-profile-wrapper.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+# If VERBOSE is true, then output line-by-line execution
+"${VERBOSE:-false}" && set -x
+"${EXIT_ON_FAILURE:-true}" && set -e
+
 # This script is a wrapper for the upgrade-profile-repo.sh script and may be used to aid the operator in updating the
 # profile repo to a target Beluga version. It abstracts away the location of the upgrade-profile-repo.sh script, which
 # performs the actual profile migration to a target Beluga version. The script must be run from the root of the